feat: add structure for favs
