+++ src/favorites/interfaces/favorite.interface.ts
@@ -1,0 +1,7 @@
+interface Favorites {
+  artists: string[]; // favorite artists ids
+  albums: string[]; // favorite albums ids
+  tracks: string[]; // favorite tracks ids
+}
+
+export { Favorites };
